Refactor init_opentelemetry
